# docs/swagger.yaml
basePath: /api
definitions:
  model.BookingResponse:
    properties:
      created_at:
        type: string
      end_time:
        type: string
      id:
        type: string
      room:
        $ref: '#/definitions/model.Room'
      room_id:
        type: string
      start_time:
        type: string
      status:
        $ref: '#/definitions/model.BookingStatus'
      updated_at:
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: string
    type: object
  model.BookingStatus:
    enum:
    - active
    - cancelled
    type: string
    x-enum-varnames:
    - BookingStatusActive
    - BookingStatusCancelled
...
  model.CreateMyBookingInput:
    properties:
      end_time:
        type: string
      room_id:
        type: string
      start_time:
        type: string
    required:
    - end_time
    - room_id
    - start_time
    type: object
  model.CreateRoomInput:
    properties:
      capacity:
        example: 10
        type: integer
      name:
        example: Meeting Room 1
        type: string
    required:
    - capacity
    - name
    type: object
  model.CreateUserInput:
    properties:
      email:
        example: riparuk@gmail.com
        type: string
      name:
        example: Rifa Faruqi
        type: string
      password:
        example: strongpassword
        type: string
      role:
        allOf:
        - $ref: '#/definitions/model.UserRole'
        example: user
    required:
    - email
    - name
    - password
    - role
    type: object
...
  model.Room:
    properties:
      capacity:
        type: integer
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  model.UpdateRoomInput:
    properties:
      capacity:
        example: 10
        type: integer
      name:
        example: Meeting Room 1
        type: string
    required:
    - capacity
    - name
    type: object
  model.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        $ref: '#/definitions/model.UserRole'
      updated_at:
        type: string
    type: object
  model.UserRole:
    enum:
    - user
    - admin
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleAdmin
info:
  contact:
    email: support@swagger.io
    name: API Support
  description: API documentation for Meet Book API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Meet Book API
  version: 1.0.0
paths:
...
  /me:
    get:
      description: Get the authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - me
  /me/bookings:
    get:
      description: Get a list of all bookings for the currently authenticated user
      parameters:
      - description: Filter by status (e.g., 'active', 'cancelled')
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of user's bookings
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/model.BookingResponse'
                type: array
            type: object
        "400":
          description: Invalid user ID
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Failed to fetch bookings
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user's bookings
      tags:
      - me
    post:
      consumes:
      - application/json
      description: Create a new room booking for the currently authenticated user
      parameters:
      - description: Booking details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.CreateMyBookingInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.BookingResponse'
      security:
      - BearerAuth: []
      summary: Create a new booking for the authenticated user
      tags:
      - me
  /rooms:
    get:
      description: Get a list of all meeting rooms
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Room'
            type: array
      summary: Get all rooms
      tags:
      - rooms
    post:
      consumes:
      - application/json
      description: Create a new meeting room
      parameters:
      - description: Room details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.CreateRoomInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Room'
      security:
      - BearerAuth: []
      summary: Create a new room
      tags:
      - rooms
  /rooms/{id}:
    delete:
      description: Delete a room by ID
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete a room
      tags:
      - rooms
    get:
      description: Get a room by its ID
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Room'
      summary: Get a room by ID
      tags:
      - rooms
    put:
      consumes:
      - application/json
      description: Update an existing room
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      - description: Room details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.UpdateRoomInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Room'
      security:
      - BearerAuth: []
      summary: Update a room
      tags:
      - rooms
  /users:
    get:
      consumes:
      - application/json
      description: Get all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: name
        in: body
        name: name
        required: true
        schema:
          $ref: '#/definitions/model.CreateUserInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.User'
      security:
      - BearerAuth: []
      summary: Create a new user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"